ChatBot-OpenAI/api/requirements.txt
langchain
langchain-openai
streamlit
python-dotenv
langchain_community
langserve
fastapi
uvicorn